DRY
https://github.com/Open-EO/openeo-geopyspark-driver/issues/920
